+++ USPAnalytics/db/migrate/20180428002617_create_orcamento_orgao_apoios.rb
@@ -7,7 +7,7 @@ class CreateOrcamentoOrgaoApoios < ActiveRecord::Migration[5.1]
       t.bigint :dotacao
       t.bigint :outros_custeios_investimentos_dotacao_basica_adicionais
       t.bigint :outros_custeios_investimentos_utilidade_publica
-      t.smallint :ano
+      t.integer :ano
 
       t.timestamps
     end
